+++ Microsoft.Azure.Cosmos/src/Query/Core/QueryResponseFactory.cs
@@ -1,0 +1,65 @@
+﻿// ------------------------------------------------------------
+// Copyright (c) Microsoft Corporation.  All rights reserved.
+// ------------------------------------------------------------
+
+namespace Microsoft.Azure.Cosmos.Query.Core
+{
+    using System;
+
+    internal static class QueryResponseFactory
+    {
+        private static readonly string EmptyGuidString = Guid.Empty.ToString();
+
+        public static QueryResponseCore CreateFromException(Exception exception)
+        {
+            QueryResponseCore queryResponseCore;
+            if (exception is CosmosException cosmosException)
+            {
+                queryResponseCore = CreateFromCosmosException(cosmosException);
+            }
+            else if (exception is Microsoft.Azure.Documents.DocumentClientException documentClientException)
+            {
+                queryResponseCore = CreateFromDocumentClientException(documentClientException);
+            }
+            else
+            {
+                // Unknown exception type should become a 500
+                queryResponseCore = QueryResponseCore.CreateFailure(
+                    statusCode: System.Net.HttpStatusCode.InternalServerError,
+                    subStatusCodes: null,
+                    errorMessage: exception.ToString(),
+                    requestCharge: 0,
+                    activityId: QueryResponseCore.EmptyGuidString,
+                    diagnostics: QueryResponseCore.EmptyDiagnostics);
+            }
+
+            return queryResponseCore;
+        }
+
+        private static QueryResponseCore CreateFromCosmosException(CosmosException cosmosException)
+        {
+            QueryResponseCore queryResponseCore = QueryResponseCore.CreateFailure(
+                statusCode: cosmosException.StatusCode,
+                subStatusCodes: (Microsoft.Azure.Documents.SubStatusCodes)cosmosException.SubStatusCode,
+                errorMessage: cosmosException.Message,
+                requestCharge: 0,
+                activityId: cosmosException.ActivityId,
+                diagnostics: QueryResponseCore.EmptyDiagnostics);
+
+            return queryResponseCore;
+        }
+
+        private static QueryResponseCore CreateFromDocumentClientException(Microsoft.Azure.Documents.DocumentClientException documentClientException)
+        {
+            QueryResponseCore queryResponseCore = QueryResponseCore.CreateFailure(
+                statusCode: documentClientException.StatusCode.GetValueOrDefault(System.Net.HttpStatusCode.InternalServerError),
+                subStatusCodes: null,
+                errorMessage: documentClientException.Message,
+                requestCharge: 0,
+                activityId: documentClientException.ActivityId,
+                diagnostics: QueryResponseCore.EmptyDiagnostics);
+
+            return queryResponseCore;
+        }
+    }
+}